==> NOTEBOOKS/part1_nlp_bert.ipynb <==
import pandas as pd
import os
import warnings

warnings.filterwarnings('ignore')

#pd.set_option('display.max_columns', None)

df = pd.read_csv("data/turkish_movie_sentiment_dataset.csv")

print(f'shape of the dataset is: {df.shape} \n')
df.head()

target = 'comment'
scores = 'point'
names  = 'film_name'

- I dont wanna remember column names everytime so I standardized with my mind like this. (Dont mind, I do this to make steps easier for me)

#### None and Duplicated Columns (check)
- Actually I'll check None values and missing inputs couple of times

print(df.isna().sum())

print(df.duplicated().sum())

df.drop_duplicates(inplace=True)

print(df.shape)

#### Taking movies with more than 30 comments
- I'll use a pretrained model, And for my case size of the dataset is not important.
I followed a way like this because movies with comment more than a x threshold'll be seen better in Django app. And With this method complications'll be reduces.

movies_w_30p_comments = df[names].value_counts()[df[names].value_counts() > 30].index

target_indexes = [ind for ind, name in enumerate(df[names]) if name in movies_w_30p_comments]

df = df.iloc[target_indexes, :].reset_index(drop=True)

df.head(10)

### Dropping comments with more than 200 words
I think long comments could be complicated and In my interface, they wont seem good

long_comments = [ind for ind, comment in enumerate(df[target]) if len(comment.split()) > 200]

print(f'Number of the problematic comments bc of size is: {len(long_comments)} ot of {len(df)}')

df = df[~df.index.isin(long_comments)].reset_index(drop=True)

print(f'len of our dataset is: {len(df)}')

### Cleaning Phase
There is some data cleaning steps are needed. Actually NLP requires some preprocessing, but beyond the NLP preprocessing steps there are some easily seen steps.
- Begining of the comments there is '\n',
- And scores in the 'point' columns arent seen as float bc of ',' (they should be '.')

# Converting point column from string to float
df[scores] = df[scores].apply(lambda x: x.replace(',', '.'))
df[scores] = df[scores].astype(float)

# We can see some \n at the begining and the ending of the comments. And there is some more spaces actually
test_comment = df[target][0]
test_comment

# Actually this one didnt solve our problem directly
test_comment.replace('\n', '')

first = test_comment.find('İlk')

last = test_comment.find('etmiş.') # length of the etmis. = 6

number_of_ending_spaces = len(test_comment) - last - 6

print(f"First {first} character is 'space' and the last {number_of_ending_spaces} character is also 'space' in our case (for first comment)")

# So in our case If I pass first 23 characters and last 22 characters of my comment, spaces are gonna be removed.

test_comment[23:-22]

- Looks good. But I have to test this for other comments (I'll show some of them)

for comment in df[target][:5]:
    print(comment[23:-22], '\n')

- Yup, everything looks good.

df[target] = df[target].apply(lambda x: x[23:-22])

df['original_comments'] = df[target].copy()

### NLP Preprocessing

import pandas as pd
from tqdm.notebook import tqdm
tqdm.pandas()

#!pip install torch===1.6.0 torchvision===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html


from transformers import AutoModelForSequenceClassification, AutoTokenizer, pipeline

model = AutoModelForSequenceClassification.from_pretrained("savasy/bert-base-turkish-sentiment-cased")
tokenizer = AutoTokenizer.from_pretrained("savasy/bert-base-turkish-sentiment-cased")

sa = pipeline("sentiment-analysis", tokenizer=tokenizer, model=model)

import re
import sys
from nltk.stem.porter import PorterStemmer

def preprocess_word(word):
    # Remove punctuation
    word = word.strip('\'"?!,.():;')
    # Convert more than 2 letter repetitions to 2 letter
    # funnnnny --> funny
    word = re.sub(r'(.)\1+', r'\1\1', word)
    return word


def preprocess_text(text):
    processed_text = []
    
    text = text.lower()

    # Replace 2+ or more dots with space
    text = re.sub(r'\.{2,}', ' ', text)
    
    # Strip space, " and '
    text = text.strip(' "\'')

    text = re.sub(r'\s+', ' ', text)
    words = text.split()

    for word in words:
      word = preprocess_word(word)
      processed_text.append(word)

    return ' '.join(processed_text)


def sentiment_analysis(text):
    "True if positive"
    p = sa(str(text))
    return pd.Series([p[0]["label"], round(p[0]["score"],2)])


df["sentiment_label"] = "-"
df[["sentiment_label", "sentiment_score"]] = df[target].progress_apply(sentiment_analysis)


df_to_save = df[[names, 'sentiment_label', 'original_comments']]

df_to_save.columns = ['movie', 'sentiment', 'comment']

df_to_save.to_csv('data/predicteds.csv', index=False)

df[names].nunique()

### Check Success Rate

gr25 = (df[scores] > 2.5).sum()
tp_gr25 = ((df[scores] > 2.5) & (df['sentiment_label'] == 'positive')).sum()

lw25 = (df[scores] < 2.5).sum()
tn_lw25 = ((df[scores] < 2.5) & (df['sentiment_label'] != 'positive')).sum()
print(f"""
Total numbers of point with greater than 2.5 are {gr25} and the # of True classified is {tp_gr25}, True classified rate: {tp_gr25 / gr25} \n
Total numbers of point with greater than 2.5 are {lw25} and the # of True classified is {tn_lw25}, True classified rate: {tn_lw25 / lw25}
""")

* Success metrics is not bed, so I wont do any additional regularization.
