==> NOTEBOOKS/part2_scrape_images.ipynb <==
I need to images of the every movies. I want to show them in my django web page with their names and the header photos.

Actually there is about 300+ movies and I dont want to search every of them by hand and save their photos. So I decided to automate this task.

I thought that I can search every movies in beyazperde.com and find their header photos, So my header photos'll fed by beyazperde.com

import requests #!pip install requests
from bs4 import BeautifulSoup #!pip install beautifulsoup4
import time

import pandas as pd

# Our When we search the name of the movie, our url should looks like this https://www.beyazperde.com/aramak/?q=kelebegin-ruyasi
BASE_SEARCHING_URL = 'https://www.beyazperde.com/aramak/?q='

def get_soup(TARGET_URL):
    page = requests.get(TARGET_URL)
    soup = BeautifulSoup(page.content, 'html.parser')
    return soup

def find_header_url(movie_name):
    movie_name = movie_name.replace(' ', '+').upper()
    SEARCHING_URL = BASE_SEARCHING_URL + movie_name
    
    searching_soup = get_soup(SEARCHING_URL)
    header_url = searching_soup.find_all('div', attrs={'class':'card entity-card entity-card-list cf'})[0].find_all('img')[0]['data-src']
    return header_url

### Taking Movie List

df = pd.read_csv('data/predicteds.csv')
df = df.dropna().reset_index(drop=True)
df.head()

movie_header_urls = {}
for name in df['movie'].unique():
    header_url = find_header_url(name)

    movie_header_urls[name] = header_url
    
    time.sleep(3)
    # I dont wanna effect traffic of the website
    # Sleep provide us to wait 3seconds in every iteration, to our all 86 requests'll send 86 * 3 = 238second

len(movie_header_urls)

df['header_urls'] = [movie_header_urls[name] for name in df['movie']]

df.head()

df.to_csv('data/predicteds_w_headers.csv', index = False)
df.head()

len(df['header_urls'][0])
