==> NOTEBOOKS/part3_insert_to_sqlite.ipynb <==
import pandas as pd
import sqlite3

from datetime import datetime, timedelta
import random
import time
import numpy as np

df = pd.read_csv('data/predicteds_w_headers.csv')

df.head()

conn = sqlite3.connect('C:/Users/bilal/Desktop/MovieApp/db.sqlite3')
c = conn.cursor()

index_ = 1

for comment, movie, sentiment, header_urls in zip(df['comment'], df['movie'], df['sentiment'], df['header_urls']):

    starting = random.randint(0,24)
    ending = random.randint(25,55)
    week = random.randint(starting, ending)
    day = random.randint(starting, ending)
    hour =  random.randint(starting, ending)
    minute = random.randint(starting, ending)

    if comment in [None, np.NaN, '']:
        comment = 'Bos Yorum'

    records = [(index_, movie, comment, sentiment, header_urls, str(datetime.now() - timedelta(weeks=week, days=day, hours=hour, minutes=minute)))]

    c.executemany('INSERT INTO moviecomments_movie VALUES(?,?,?,?,?,?);',records)

    index_ += 1

    if index_ % 10 == 0:
        time.sleep(0.2)
			
conn.commit()
conn.close()
